+++ src/main/java/com/kh/bts/investment/model/vo/Sold.java
@@ -6,14 +6,14 @@ import org.springframework.stereotype.Component;
 public class Sold {
 	private int usno;
 	private String coin;
-	private float sellcnt;
-	private float sellprice;
+	private double sellcnt;
+	private double sellprice;
 	private String sdate;
 	private String acntno;
 
 	public Sold() {
 	}
-	public Sold(int usno,String coin,float sellcnt,float sellprice,String sdate,String acntno) {
+	public Sold(int usno,String coin,double sellcnt,double sellprice,String sdate,String acntno) {
 		super();
 		this.usno = usno;
 		this.coin = coin;
@@ -39,19 +39,19 @@ public class Sold {
 		this.coin = coin;
 	}
 
-	public float getSellcnt() {
+	public double getSellcnt() {
 		return sellcnt;
 	}
 
-	public void setSellcnt(float sellcnt) {
+	public void setSellcnt(double sellcnt) {
 		this.sellcnt = sellcnt;
 	}
 
-	public float getSellprice() {
+	public double getSellprice() {
 		return sellprice;
 	}
 
-	public void setSellprice(float sellprice) {
+	public void setSellprice(double sellprice) {
 		this.sellprice = sellprice;
 	}
